Added New Game Session by button. Clean up Unused components Reorganise file structure content.
