--- src/lib/components/CheckmarkButton.svelte
@@ -1,8 +1,0 @@
-<!-- src/lib/components/CheckmarkButton.svelte -->
-<div
-	class="absolute top-5 right-5 flex h-[30px] w-[30px] cursor-pointer items-center justify-center rounded-full bg-gray-800"
->
-	<svg class="h-4 w-4 stroke-white stroke-2" viewBox="0 0 24 24">
-		<polyline points="20 6 9 17 4 12"></polyline>
-	</svg>
-</div>
