--- src/lib/components/GameAddButton.svelte
@@ -1,15 +1,0 @@
-<!-- src/lib/components/GameAddButton.svelte -->
-<button
-	class="flex items-center justify-center rounded bg-gray-800 px-2 py-1 text-sm hover:bg-gray-700 md:px-4 md:py-2 md:text-base"
->
-	<svg
-		class="mr-1 h-2.5 w-2.5 stroke-current md:mr-1.5 md:h-5 md:w-5"
-		viewBox="0 0 24 24"
-		fill="none"
-		stroke-width="2"
-	>
-		<rect x="3" y="3" width="18" height="18" rx="2" ry="2"></rect>
-		<path d="M3 3 h18 v18 h-18 v-18 M8 12 h8 M12 8 v8"></path>
-	</svg>
-	Game
-</button>
