--- src/lib/components/HintButton.svelte
@@ -1,16 +1,0 @@
-<!-- src/lib/components/HintButton.svelte -->
-<button
-	class="flex items-center justify-center rounded bg-gray-800 px-2 py-1 text-sm hover:bg-gray-700 md:px-4 md:py-2 md:text-base"
->
-	<svg
-		class="mr-1 h-2.5 w-2.5 stroke-current md:mr-1.5 md:h-5 md:w-5"
-		viewBox="0 0 24 24"
-		fill="none"
-		stroke-width="2"
-	>
-		<path
-			d="M12 2 a10 10 0 1 0 0 20 a10 10 0 0 0 0 -20 M9.09 9 a3 3 0 0 1 5.83 1 c0 2-3 3-3 3 M12 17 v0.01"
-		></path>
-	</svg>
-	Hint
-</button>
