--- src/lib/components/StarButton.svelte
@@ -1,10 +1,0 @@
-<!-- src/lib/components/StarrkButton.svelte -->
-<div
-	class="absolute top-16 right-5 flex h-[30px] w-[30px] cursor-pointer items-center justify-center rounded-full bg-gray-800"
->
-	<svg class="h-4 w-4 fill-none stroke-white stroke-2" viewBox="0 0 24 24">
-		<polygon
-			points="12 2 15.09 8.26 22 9.27 17 14.14 18.18 21.02 12 17.77 5.82 21.02 7 14.14 2 9.27 8.91 8.26 12 2"
-		></polygon>
-	</svg>
-</div>
