+++ src/lib/constants.ts
@@ -1,5 +1,3 @@
-export const CARD_SIZE_ASPECT_RATIO = 1.57; // (w x h : 7x11)
-
 export const RESPONSIVE_CLASS = {
 	CARD_SIZE:
 		'h-11 w-7 @sm:h-16.5 @sm:w-9.5 @xl:h-22 @xl:w-14 @3xl:h-24.75 @3xl:w-15.75 @5xl:h-33 @5xl:w-21 @7xl:h-44 @7xl:w-32',
@@ -10,4 +8,27 @@ export const RESPONSIVE_CLASS = {
 	GAP_SIZE: 'gap-0.5 @xl:gap-1 @3xl:gap-1.5 @5xl:gap-2',
 	// PADDING_SIZE: 'p-0.5 @xl:p-1 @3xl:p-1.5 @5xl:p-2'
 	PADDING_SIZE: 'p-0.5 @2xl:p-1 @7x:p-2'
+} as const;
+
+export const CARD_VALUES = [
+	'A',
+	'2',
+	'3',
+	'4',
+	'5',
+	'6',
+	'7',
+	'8',
+	'9',
+	'10',
+	'J',
+	'Q',
+	'K'
+] as const;
+
+export const NO_OF_CARD_SLOT = 10;
+
+export const CARD_COLORS = {
+	RED: 'text-red-500',
+	BLACK: 'text-black'
 } as const;
